+++ src/components/badges/license-badge.tsx
@@ -4,17 +4,25 @@ import { Badge, BadgeProps } from "@chakra-ui/react";
 function LicenseBadge({
   letter,
   color,
+  themeInverted,
   ...rest
 }: BadgeProps & {
   letter: string;
   color: string;
+  themeInverted?: boolean;
 }) {
   const dark = darkenHexColor(`#${color}`, 0.3);
   const light = lightenHexColor(`#${color}`, 0.7);
   return (
     <Badge
-      bg={{ base: light, _dark: dark }}
-      color={{ base: dark, _dark: light }}
+      bg={{
+        base: !themeInverted ? light : dark,
+        _dark: !themeInverted ? dark : light,
+      }}
+      color={{
+        base: !themeInverted ? dark : light,
+        _dark: !themeInverted ? light : dark,
+      }}
       borderColor={`#${color}`}
       borderWidth={1}
       fontWeight={"bold"}